requirements.txt
Django==1.8.5
django-admin-bootstrapped==2.5.7
django-wysiwyg-redactor==0.4.9.1
